+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [10.0.0] - 2020-09-02
+
+### Changed
+
+-   Changed [Select](src/components/select/select.tsx) API
+-   Add [Select](src/components/select/select.tsx) support in native and menu mode
+-   Add [Option](src/components/select/option.tsx)
+-   Add [OptionLabel](src/components/select/option-label.tsx)
+-   Add [SelectEmitter](src/components/select/select-emitter.tsx)
+-   Add [ArrowBottomThinIcon](src/icons/arrow-bottom-thin-icon.tsx)
+
 ## [9.0.3] - 2020-09-15
 
 ### Added
@@ -35,7 +46,7 @@
 
 ### Changed
 
--   Chnaged `<Paper />` to consume dark context and change own appearance depending on its value
+-   Changed `<Paper />` to consume dark context and change own appearance depending on its value
 -   Updated scrollbar colors in `<ScrollView />`
 
 ## [8.4.1] - 2020-06-30